remove things unrelated to lvlscript out of lvl_script.h
remove things unrelated to lvlscript out of lvl_script.h and move them to a more appropriate location including new file map_locations
