+++ src/actionpt.h
@@ -65,6 +65,8 @@ TbBool action_point_is_invalid(const struct ActionPoint *apt);
 
 TbBool action_point_reset_idx(ActionPointId apt_idx);
 unsigned long get_action_point_activated_by_players_mask(ActionPointId apt_idx);
+TbBool action_point_activated_by_player(ActionPointId apt_idx, PlayerNumber plyr_idx);
+
 void clear_action_points(void);
 void delete_all_action_point_structures(void);
 TbBool process_action_points(void);
