+++ src/creature_groups.c
@@ -855,6 +855,23 @@ struct Thing *script_process_new_party(struct Party *party, PlayerNumber plyr_id
     return leadtng;
 }
 
+void script_process_new_tunneller_party(PlayerNumber plyr_idx, long prty_id, TbMapLocation location, TbMapLocation heading, unsigned char crtr_level, unsigned long carried_gold)
+{
+    struct Thing* ldthing = script_process_new_tunneler(plyr_idx, location, heading, crtr_level, carried_gold);
+    if (thing_is_invalid(ldthing))
+    {
+        ERRORLOG("Couldn't create tunneling group leader");
+        return;
+    }
+    struct Thing* gpthing = script_process_new_party(&gameadd.script.creature_partys[prty_id], plyr_idx, location, 1);
+    if (thing_is_invalid(gpthing))
+    {
+        ERRORLOG("Couldn't create creature group");
+        return;
+    }
+    add_creature_to_group_as_leader(ldthing, gpthing);
+}
+
 /******************************************************************************/
 #ifdef __cplusplus
 }
