+++ src/creature_states_hero.h
@@ -21,6 +21,7 @@
 
 #include "bflib_basics.h"
 #include "globals.h"
+#include "map_locations.h"
 
 #ifdef __cplusplus
 extern "C" {
@@ -63,10 +64,7 @@ short setup_person_tunnel_to_position(struct Thing *creatng, MapSubtlCoord stl_x
 long get_best_dungeon_to_tunnel_to(struct Thing *creatng);
 TbBool send_tunneller_to_point_in_dungeon(struct Thing *creatng, PlayerNumber plyr_idx, struct Coord3d *pos);
 TbBool is_hero_tunnelling_to_attack(struct Thing *creatng);
-TbBool script_support_send_tunneller_to_action_point(struct Thing *thing, long apt_idx);
-TbBool script_support_send_tunneller_to_dungeon(struct Thing *creatng, PlayerNumber plyr_idx);
-TbBool script_support_send_tunneller_to_dungeon_heart(struct Thing *creatng, PlayerNumber plyr_idx);
-TbBool script_support_send_tunneller_to_appropriate_dungeon(struct Thing *creatng);
+struct Thing *script_process_new_tunneler(unsigned char plyr_idx, TbMapLocation location, TbMapLocation heading, unsigned char crtr_level, unsigned long carried_gold);
 /******************************************************************************/
 #ifdef __cplusplus
 }
