+++ src/lvl_script_lib.c
@@ -137,21 +137,6 @@ TbBool get_map_location_id_f(const char *locname, TbMapLocation *location, const
     return true;
 }
 
-void command_add_value(unsigned long var_index, unsigned long plr_range_id, long val2, long val3, long val4)
-{
-    ALLOCATE_SCRIPT_VALUE(var_index, plr_range_id);
-
-    value->arg0 = val2;
-    value->arg1 = val3;
-    value->arg2 = val4;
-
-    if ((get_script_current_condition() == CONDITION_ALWAYS) && (next_command_reusable == 0))
-    {
-        script_process_value(var_index, plr_range_id, val2, val3, val4, value);
-        return;
-    }
-}
-
 long parse_creature_name(const char *creature_name)
 {
     long ret = get_rid(creature_desc, creature_name);
@@ -165,72 +150,7 @@ long parse_creature_name(const char *creature_name)
     return ret;
 }
 
-TbBool get_coords_at_hero_door(struct Coord3d *pos, long gate_num, unsigned char random_factor)
-{
-    SYNCDBG(7,"Starting at HG%d", (int)gate_num);
-    if (gate_num <= 0)
-    {
-        ERRORLOG("Script error - invalid hero gate index %d",(int)gate_num);
-        return false;
-    }
-    struct Thing* gatetng = find_hero_gate_of_number(gate_num);
-    if (thing_is_invalid(gatetng))
-    {
-        ERRORLOG("Script error - attempt to create thing at non-existing hero gate index %d",(int)gate_num);
-        return false;
-    }
-    pos->x.val = gatetng->mappos.x.val;
-    pos->y.val = gatetng->mappos.y.val;
-    pos->z.val = gatetng->mappos.z.val + 384;
-    return true;
-}
 
-TbBool get_coords_at_action_point(struct Coord3d *pos, long apt_idx, unsigned char random_factor)
-{
-    SYNCDBG(7,"Starting at action point %d", (int)apt_idx);
-
-    struct ActionPoint* apt = action_point_get(apt_idx);
-    if (!action_point_exists(apt))
-    {
-        ERRORLOG("Script error - attempt to create thing at non-existing action point %d",(int)apt_idx);
-        return false;
-    }
-
-    if ( (random_factor == 0) || (apt->range == 0) )
-    {
-        pos->x.val = apt->mappos.x.val;
-        pos->y.val = apt->mappos.y.val;
-    } else
-    {
-        long direction = GAME_RANDOM(2 * LbFPMath_PI);
-        long delta_x = (apt->range * LbSinL(direction) >> 8);
-        long delta_y = (apt->range * LbCosL(direction) >> 8);
-        pos->x.val = apt->mappos.x.val + (delta_x >> 8);
-        pos->y.val = apt->mappos.y.val - (delta_y >> 8);
-    }
-    return true;
-}
-
-/**
- * Creates a thing on given players dungeon heart.
- * Originally was script_support_create_creature_at_dungeon_heart().
- * @param plyr_idx
- */
-TbBool get_coords_at_dungeon_heart(struct Coord3d *pos, PlayerNumber plyr_idx)
-{
-    SYNCDBG(7,"Starting at player %d", (int)plyr_idx);
-    struct Thing* heartng = get_player_soul_container(plyr_idx);
-    TRACE_THING(heartng);
-    if (thing_is_invalid(heartng))
-    {
-        ERRORLOG("Script error - attempt to create thing in player %d dungeon with no heart",(int)plyr_idx);
-        return false;
-    }
-    pos->x.val = heartng->mappos.x.val + PLAYER_RANDOM(plyr_idx, 65) - 32;
-    pos->y.val = heartng->mappos.y.val + PLAYER_RANDOM(plyr_idx, 65) - 32;
-    pos->z.val = heartng->mappos.z.val;
-    return true;
-}
 
 struct Thing *script_process_new_object(long tngmodel, TbMapLocation location, long arg)
 {
